Pre-JWT: req.body
JSON web tokens are required because req.session doesn't persist like it does on the backend project. In that project, everything was rendered from the backend and we never really left it. Now, req.session is cleared with every new fetch request. JWT allow us to communicate things like which user is logged in and which room they're in with the backend, which is required for join table management. Right now, we'll use req.body to communicate these things instead of JWTs.